# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Haraj Station API
  version: 1.0.0
  description: Generated inventory of route handlers. Request and response schemas are being added per mobile contract.
servers:
  - url: https://harajstation.com
paths:
...
  "/api/auctions/{id}":
    get:
      summary: "GET /api/auctions/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
  "/api/auctions/{id}/bids":
    post:
      summary: "POST /api/auctions/{id}/bids"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/auth/login":
    post:
      summary: "POST /api/auth/login"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/cron":
    get:
      summary: "GET /api/cron"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/identity/doc/{id}":
    get:
      summary: "GET /api/identity/doc/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...